Tested with jest add abd remove item
