+++ modules/storageLocal.js
@@ -31,6 +31,23 @@ export default class StorageLocal {
        localStorage.setItem('todos', JSON.stringify(todos));
        this.index += 1;
      };
+
+    removeTodo = (index) => {
+      let todos = JSON.parse(localStorage.getItem('todos'));
+      if (todos.length === 1) {
+        todos = [];
+        localStorage.setItem('todos', JSON.stringify(todos));
+        return;
+      }
+
+      todos = todos.filter((todo) => todo.index !== parseInt(index, 10));
+      localStorage.setItem('todos', JSON.stringify(todos));
+
+      todos.forEach((todo, i) => {
+        todo.index = i + 1;
+      });
+      localStorage.setItem('todos', JSON.stringify(todos));
+    };
 }
 
 export const store = new StorageLocal();
