+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "private": true,
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest --coverage",
     "build": "webpack",
     "start": "webpack serve --open",
     "server": "node server.js",
@@ -22,7 +22,10 @@
   },
   "homepage": "https://github.com/AlexisJW/to-do-list-structure",
   "devDependencies": {
+    "@babel/core": "^7.20.12",
+    "@babel/preset-env": "^7.20.2",
     "babel-eslint": "^10.1.0",
+    "babel-jest": "^29.4.1",
     "css-loader": "^6.7.3",
     "csv-loader": "^3.0.5",
     "eslint": "^7.32.0",
@@ -31,6 +34,8 @@
     "gh-pages": "^5.0.0",
     "hint": "^7.1.3",
     "html-webpack-plugin": "^5.5.0",
+    "jest": "^29.4.1",
+    "jest-environment-jsdom": "^29.4.1",
     "json5": "^2.2.3",
     "style-loader": "^3.3.1",
     "stylelint": "^13.13.1",
@@ -44,8 +49,8 @@
     "yamljs": "^0.3.0"
   },
   "dependencies": {
-    "lodash": "^4.17.21",
     "express": "^4.18.2",
+    "lodash": "^4.17.21",
     "webpack": "^5.75.0",
     "webpack-dev-middleware": "^6.0.1"
   }
